refactor: isolate babel-types patching
