+++ packages/babel-types/src/types/definitions.js
@@ -13,7 +13,7 @@ const valueFieldCommon = {
   }
 };
 
-export default {
+const MarkoDefinitions = {
   MarkoDocumentType: {
     aliases: ["Marko", "Statement"],
     builder: ["value"],
@@ -204,3 +204,9 @@ export default {
     }
   }
 };
+
+export default MarkoDefinitions;
+export const MARKO_TYPES = Object.keys(MarkoDefinitions);
+export const MARKO_ALIAS_TYPES = Array.from(
+  new Set(MARKO_TYPES.flatMap(t => MarkoDefinitions[t].aliases))
+);
